pyproject.toml
[project]
name = "job-application-tracker"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135.3",
    "hashward[bcrypt]>=0.1.0",
    "psycopg2>=2.9.11",
    "pydantic[email]>=2.12.5",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "python-jose[cryptography]>=3.5.0",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.44.0",
    "python-multipart>=0.0.26",
    "alembic>=1.18.4",
    "google-api-python-client>=2.194.0",
    "google-auth>=2.49.2",
    "google-auth-oauthlib>=1.3.1",
    "jinja2>=3.1.6",
    "redis>=7.4.0",
    "celery>=5.6.3",
]
